+++ doc/general/changelog.rst
@@ -14,6 +14,9 @@ Change from v0.3 to TODO
 * renaming of ``bip.base.Operand`` to ``bip.base.BipOperand``
 * renaming of ``bip.hexrays.HexRaysEvent`` to ``bip.base.HxEvent``
 * renaming of ``bip.base.Instr`` to ``bip.base.BipInstr``
+* renaming of ``bip.base.BipFuncFlags`` to ``bip.base._BipFuncFlags``
+* renaming of ``bip.base.BipFlowChartFlag`` to ``bip.base._BipFlowChartFlag``
+
 
 
 Sed script for automatic update of plugins (no garantee to be perfect or to
@@ -28,5 +31,6 @@ avoid colisions) (use with ``sed -f RULEFILE INPUTFILE``):
     s/countBipOperand/countOperand/g
     s/HexRaysEvent/HxEvent/g
     s/Instr/BipInstr/g
-
+    s/BipFuncFlags/_BipFuncFlags/g
+    s/BipFlowChartFlag/_BipFlowChartFlag/g